Refactor Sleep Entries also
